foodanddrink_version2.0.html
This is the homepage of the second edition of the food module that I designed. Based on the first edition, I made the following changes: Wen Yang and I modified the navigation bar after discussion, and added some interactive effects. In addition, the web content uses flexbox layout to achieve the content to adapt to the window changes! If you have any suggestions, please leave a comment here.